ml-service/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
debugpy==1.8.7
Django==5.1.2
django-cors-headers==4.5.0
django-rest-framework==0.1.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
protobuf==5.28.3
psycopg2==2.9.10
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.16.0
